% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplot.R
\name{corplot}
\alias{corplot}
\title{Correlation plot}
\usage{
corplot(cor_mat, absolute = TRUE, var_range = 1:ncol(cor_mat))
}
\arguments{
\item{cor_mat}{Any squared matrix (usually a correlation matrix.)}
}
\value{
A ggplot object containing the heatmap.
}
\description{
Create a correlation heatmap for explaining the performance of MI-PCA.
}
\examples{
# Crete an example correlation matrix
cor_mat <- cor(mtcars)

# use it
corplot(cor_mat)

# Check if you want absolute values or not
cor_mat <- matrix(c(1, -.5, -.5, 1), ncol = 2, dimnames = list(c("V1", "V2"), c("V1", "V2")))

# use it
corplot(cor_mat)

}
\author{
Edoardo Costantini, 2023
}
