% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Function full name}
\usage{
gen_data(
  n = 1000,
  block_sizes = c(5, 5, 10),
  block_rhos = c(0.6, 0.3, 0.01),
  inflated_rhos = NULL,
  inflated_values = NULL
)
}
\arguments{
\item{n}{A description of the first argument}

\item{p}{A description of the second argument}
}
\value{
Description of function output
}
\description{
A high level description of the function.
}
\details{
This function does such and such.
}
\examples{

# Example internals
n <- 1e3
block_sizes <- c(5, 5, 10)
block_rhos <- c(.6, .3, .01)
inflated_rhos <- list(c(4, 5), c(9, 10), c(11:p))
inflated_values <- .9

# Specification that reproduces results from the simulation study in Costantini et al 2022

gen_data(
    n = 200,
    block_sizes = c(5, 5, 50),
    block_rhos = c(.6, .3, .01),
    inflated_rhos = list(c(4, 5), c(9, 10), c(11:p)),
    inflated_values = ''.9
)

}
\references{
Costantini, E., Lang, K. M., Reeskens, T., & Sijtsma, K. (2022). High-dimensional imputation for the social sciences: a comparison of state-of-the-art methods. arXiv preprint arXiv:2208.13656.
}
\author{
Edoardo Costantini, 2023
}
