% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_correlation.R
\name{heatmap_correlation}
\alias{heatmap_correlation}
\title{Heatmap correlation}
\usage{
heatmap_correlation(cor_mat, absolute = TRUE, var_range = 1:ncol(cor_mat))
}
\arguments{
\item{cor_mat}{any squared matrix (usually a correlation matrix.)}
}
\value{
A ggplot object containing the heatmap.
}
\description{
Creates a heatmap of correlation coefficients.
}
\examples{
# Crete an example correlation matrix
cor_mat <- cor(mtcars)

# use it
heatmap_correlation(cor_mat)

# Check if you want absolute values or not
cor_mat <- matrix(c(1, -.5, -.5, 1), ncol = 2, dimnames = list(c("V1", "V2"), c("V1", "V2")))

# use it
heatmap_correlation(cor_mat)

}
\author{
Anonymized, 2023
}
