% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_loadings.R
\name{heatmap_loadings}
\alias{heatmap_loadings}
\title{Heatmap loadings}
\usage{
heatmap_loadings(
  load_mat,
  absolute = TRUE,
  var_range = 1:nrow(load_mat),
  PCs_range = 1:ncol(load_mat)
)
}
\arguments{
\item{load_mat}{a matrix of principal component loadings of dimensionality p x q (with p = number of variables and q = number of PCs)}

\item{absolute}{a logical value describing whether the absolute values of \code{load_mat} should be considered}

\item{var_range}{range of variables for which to plot the loadings}

\item{PCs_range}{range of PCs for which to plot the loadings}
}
\value{
Description of function output
}
\description{
Creates a heatmap of the principal component loadings for a PCA analysis.
}
\examples{
# Perform SVD of some data
svd_mtcars <- svd(scale(mtcars))

# Crete an example loadings matrix
load_mat <- svd_mtcars$v

# Give meaningful names to dimensions
colnames(load_mat) <- paste0("PC", 1:ncol(mtcars))
rownames(load_mat) <- colnames(mtcars)

# use it
heatmap_loadings(load_mat)

}
\author{
Edoardo Costantini, 2023
}
