% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_npcs.R
\name{histogram_npcs}
\alias{histogram_npcs}
\title{Histogram Number of Principal Components}
\usage{
histogram_npcs(npcs_nscree, npcs_50rule)
}
\arguments{
\item{npcs_nscree}{data.frame of nScree results}

\item{npcs_50rule}{number of PCs kept by 50$ rule}
}
\value{
A ggplot object containing the histogram.
}
\description{
Creates a histogram of number of components selected based on different decision rules.
}
\details{
This function does such and such.
}
\examples{
# Crete an example correlation matrix
npcs_nscree <- nFactors::nScree(as.data.frame(mtcars))$Components
npcs_50rule <- 5

# use it
histogram_npcs(npcs_nscree, npcs_50rule)

}
\author{
Anonymized, 2023
}
