% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resampling.R
\name{plot_resampling}
\alias{plot_resampling}
\title{Plot resampling study}
\usage{
plot_resampling(
  res,
  outcome = c("bias_per", "ci_cov", "CIW")[1],
  model = c("m1", "m2")[1],
  dt_reps = 500,
  ci_lvl = 0.95,
  meth_compare
)
}
\arguments{
\item{argument_1}{A description of the first argument}

\item{argument_2}{A description of the second argument}
}
\value{
Description of function output
}
\description{
Plot resampling study
}
\details{
This function does such and such.
}
\examples{
# Example input
res <- res_exp_4
model <- c("m1", "m2")[1]
dt_reps <- 500
ci_lvl <- .95
meth_compare <- c("DURR_la", "IURR_la", "blasso", "bridge", "MI_PCA", "MI_CART", "MI_RF", "stepFor", "CC")
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
