% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot results for the simulation study}
\usage{
plot_simulation(res, dims, outcome, meths, prop_NA, x_lims)
}
\arguments{
\item{res}{result object from the simulation study}

\item{dims}{number of variables used in the data generation procedure}

\item{outcome}{type of performance measure to be plotted}

\item{meths}{names of imputation methods to plot}

\item{prop_NA}{vector of the correlation coefficients to plot}

\item{x_lims}{minimum and maxim values for the x axis values}
}
\value{
ggplot object
}
\description{
Plot results for the simulation study
}
\examples{
# Example internals
data(res_exp_1)
res <- res_exp_1
dims <- 50
outcome <- c("PRB", "CIC", "CIW")[1]
meths <- levels(res$methods)[1:11]
prop_NA <- c(0.1, .3)
x_lims <- c(0, 50)

}
\author{
Edoardo Costantini, 2023
}
