% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation_colli.R
\name{plot_simulation_colli}
\alias{plot_simulation_colli}
\title{Plot results for the simulation study on collinearity}
\usage{
plot_simulation_colli(res, dims, outcome, meths, rho, x_lims)
}
\arguments{
\item{res}{result object from the simulation study 1.2 on collinearity}

\item{dims}{number of variables used in the data generation procedure}

\item{outcome}{type of performance measure to be plotted}

\item{meths}{names of imputation methods to plot}

\item{rho}{vector of the correlation coefficients to plot}

\item{x_lims}{minimum and maxim values for the x axis values}
}
\value{
ggplot object
}
\description{
Plot results for the simulation study on collinearity
}
\examples{
# Example internals
data(res_exp_1_2)
res <- res_exp_1_2
dims <- 50
outcome <- c("PRB", "CIC", "CIW")[3]
meths <- levels(res$methods)[1:11]
rho <- c(0, .9)
x_lims <- c(50, 100)

}
\author{
Edoardo Costantini, 2023
}
