% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_simulation.R
\name{plot_time_simulation}
\alias{plot_time_simulation}
\title{Plot time for different data results}
\usage{
plot_time_simulation(res, meths, dims, prop_NA, x_lims)
}
\arguments{
\item{res}{result object from the simulation study}

\item{meths}{names of imputation methods to plot}

\item{dims}{number of variables used in the data generation procedure}

\item{prop_NA}{vector of the correlation coefficients to plot}

\item{x_lims}{minimum and maxim values for the x axis values}
}
\value{
ggplot object
}
\description{
Plot time for different data results
}
\examples{
res <- res_exp_1_time
meths <- levels(dt$variable)[1:3]
dims <- 50
prop_NA <- .3
x_lims <- c(0, 90)

}
\author{
Edoardo Costantini, 2023
}
