% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi-pca.R
\name{plotResults}
\alias{plotResults}
\title{Shiny app to understand MI-PCA results}
\usage{
plotResults()
}
\arguments{
\item{argument_1}{A description of the first argument}

\item{argument_2}{A description of the second argument}
}
\value{
Starts a shiny app to explore the results.
}
\description{
This function allows you to generate data according to different conditions explored in the reference study, and explore how PCA behaves on the data.
}
\details{
This function does such and such.
}
\references{
Costantini, E., Lang, K. M., Reeskens, T., & Sijtsma, K. (2022). High-dimensional imputation for the social sciences: a comparison of state-of-the-art methods. arXiv preprint arXiv:2208.13656.
}
\author{
Edoardo Costantini, 2023
}
