% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_exp_1}
\alias{res_exp_1}
\title{res_exp_1}
\format{
A data frame
}
\usage{
data(res_exp_1)
}
\description{
The data.frame containing the results for the main simulation study.
}
\keyword{datasets}
