% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_exp_1_2}
\alias{res_exp_1_2}
\title{res_exp_1_2}
\format{
A data frame
}
\usage{
data(res_exp_1_2)
}
\description{
The data.frame containing the results fir simulation study 1.2 on collinearity.
}
\keyword{datasets}
