% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_exp_1_time}
\alias{res_exp_1_time}
\title{res_exp_1_time}
\format{
A data frame
}
\usage{
data(res_exp_1_time)
}
\description{
The data.frame containing the time results for the main simulation study.
}
\keyword{datasets}
