% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_cpve.R
\name{scatter_cpve}
\alias{scatter_cpve}
\title{Scatter plot of the CPVE for subsequent PCs}
\usage{
scatter_cpve(cpve, PCs_range = 1:length(cpve))
}
\arguments{
\item{cpve}{vector of cumulative proportion of variance explained values}
}
\value{
A ggplot object containing the scatter plot.
}
\description{
Scatter plot of the cumulative proportion of variance explained (CPVE) by principal components (PCS)
}
\details{
This function does such and such.
}
\examples{
# Perform SVD of some data
svd_mtcars <- svd(scale(mtcars))

# Compute CPVE
cpve <- cumsum(prop.table(svd_mtcars$d^2))

# use it
scatter_cpve(cpve)

}
\author{
Edoardo Costantini, 2023
}
