% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{set of inputs coming from ui}

\item{output}{set of outputs}

\item{session}{session info and status}
}
\description{
server function for the shiny app
}
\author{
Anonymized, 2023
}
