% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Start Shiny app}
\usage{
start_app()
}
\value{
Starts a shiny app to explore the results.
}
\description{
This function starts a Shiny app used to display the results from Anonymized et. al. (2022).
}
\references{
Anonymized
}
\author{
Anonymized, 2023
}
