% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_mi_pca_text.R
\name{tab_mi_pca_text}
\alias{tab_mi_pca_text}
\title{Tab text: MI-PCA}
\usage{
tab_mi_pca_text(output)
}
\arguments{
\item{output}{list of output objects used in the server}
}
\value{
Updated \code{output} object to be processed by ui and server
}
\description{
A function to add text to the server output object for the MI-PCA tab.
}
\author{
Anonymized, 2023
}
