% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_call}
\alias{ui_call}
\title{User interface call}
\usage{
ui_call()
}
\value{
UI object that can be passed directly to shiny::shinyApp()
}
\description{
Calls the definition of the user interface and returns it as an output
}
\author{
Anonymized, 2023
}
