% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot results for the simulation study}
\usage{
plot_simulation(
  res,
  dims,
  outcome,
  meths,
  rho = NULL,
  prop_NA = NULL,
  x_lims,
  reps
)
}
\arguments{
\item{res}{result object from the simulation study}

\item{dims}{number of variables used in the data generation procedure}

\item{outcome}{type of performance measure to be plotted}

\item{meths}{names of imputation methods to plot}

\item{rho}{vector of the correlation coefficients to plot}

\item{prop_NA}{vector of the correlation coefficients to plot}

\item{x_lims}{minimum and maxim values for the x axis values}

\item{reps}{number of repetitions used in the simulation}
}
\value{
ggplot object
}
\description{
Plot results for the simulation study
}
\details{
If you use the data for the first simulation study, then set \code{rho = NA}. If you use data for simulation study 1.2, then set \code{prop_NA = 0.3}.
}
\examples{
# Example internals main simulation study
data(res_exp_1)
res <- res_exp_1
dims <- 50
outcome <- c("PRB", "CIC", "CIW")[1]
meths <- levels(res_exp_1$methods)[1:11]
prop_NA <- c(0.1, .3)
rho <- unique(res_exp_1$collinearity)

# Example use
plot_simulation(
    res = res_exp_1,
    dims = dims,
    outcome = outcome,
    meths = meths,
    x_lims = x_lims,
    rho = unique(res_exp_1$collinearity),
    prop_NA = prop_NA
)

# Example internals collinearity study
data(res_exp_1_2)
res <- res_exp_1_2
prop_NA <- unique(res_exp_1_2$pm)
rho <- c(0, .9)
x_lims <- c(1, 50)

# Example use
plot_simulation(
    res = res_exp_1_2,
    dims = unique(res_exp_1_2$p)[1],
    outcome = levels(res_exp_1_2$analysis)[1],
    meths = levels(res_exp_1_2$methods)[1:5],
    x_lims = c(1, 50),
    rho = unique(res_exp_1_2$collinearity),
    prop_NA = unique(res_exp_1_2$pm)
)

}
\author{
Anonymized, 2023
}
