% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Start Shiny app}
\usage{
start_app()
}
\value{
Starts a shiny app to explore the results.
}
\description{
This function starts a Shiny app used to display the results from Costantini et. al. (2022).
}
\references{
Costantini, E., Lang, K. M., Reeskens, T., & Sijtsma, K. (2022). High-dimensional imputation for the social sciences: a comparison of state-of-the-art methods. arXiv preprint arXiv:2208.13656.
}
\author{
Edoardo Costantini, 2023
}
